RavenDB-17859 - Subscription test inside the Studio does not display results when query contains revisions
